:recycle: [maykinmedia/open-api-framework#159] refactor uml directive
