+++ .readthedocs.yml
@@ -1,5 +1,4 @@
 ---
-
 # .readthedocs.yaml
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
@@ -15,9 +14,10 @@ build:
   apt_packages:
     - libgdal-dev
     - gdal-bin
+    - graphviz
+    - graphviz-dev
     - build-essential
     - python3-dev
-
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -25,4 +25,4 @@ sphinx:
 
 python:
   install:
-  - requirements: requirements/dev.txt
+    - requirements: requirements/dev.txt
