+++ requirements/base.in
@@ -1,5 +1,5 @@
 open-api-framework[celery,cors,markup,geo,csp,commonground,inclusions,sanitization,server,redis]
-
+django-csp==3.8.0
 # Core python libraries
 jq
 jsonschema
