+++ requirements/base.txt
@@ -25,8 +25,10 @@ attrs==21.4.0
     #   requests-cache
 billiard==4.2.1
     # via celery
-bleach==6.1.0
-    # via open-api-framework
+bleach==6.2.0
+    # via
+    #   django-markup
+    #   open-api-framework
 boltons==24.0.0
     # via
     #   face
@@ -84,8 +86,6 @@ cryptography==44.0.1
     #   mozilla-django-oidc
     #   pyopenssl
     #   webauthn
-deprecated==1.2.14
-    # via redis
 dictdiffer==0.9.0
     # via -r requirements/base.in
 django==5.2.3
@@ -123,6 +123,7 @@ django==5.2.3
     #   django-two-factor-auth
     #   django-upgrade-check
     #   djangorestframework
+    #   djangorestframework-gis
     #   djangorestframework-inclusions
     #   drc-cmis
     #   drf-extra-fields
@@ -146,10 +147,12 @@ django-better-admin-arrayfield==1.4.2
     # via -r requirements/base.in
 django-celery-beat==2.8.0
     # via -r requirements/base.in
-django-cors-headers==4.4.0
+django-cors-headers==4.7.0
     # via open-api-framework
 django-csp==3.8
-    # via open-api-framework
+    # via
+    #   -r requirements/base.in
+    #   open-api-framework
 django-db-logger==0.1.12
     # via -r requirements/base.in
 django-extensions==4.1
@@ -160,6 +163,7 @@ django-filter==24.2
     # via
     #   commonground-api-common
     #   django-loose-fk
+    #   djangorestframework-gis
     #   open-api-framework
 django-formtools==2.5.1
     # via django-two-factor-auth
@@ -173,7 +177,7 @@ django-log-outgoing-requests==0.6.1
     # via open-api-framework
 django-loose-fk==1.1.2
     # via -r requirements/base.in
-django-markup==1.8.1
+django-markup==1.9.1
     # via open-api-framework
 django-ordered-model==3.7.4
     # via django-admin-index
@@ -185,7 +189,7 @@ django-privates==3.1.1
     # via
     #   -r requirements/base.in
     #   django-simple-certmanager
-django-redis==5.4.0
+django-redis==6.0.0
     # via open-api-framework
 django-relativedelta==2.0.0
     # via
@@ -236,12 +240,15 @@ djangorestframework-camel-case==1.4.2
     # via
     #   commonground-api-common
     #   notifications-api-common
-djangorestframework-gis==1.0
+djangorestframework-gis==1.2.0
     # via open-api-framework
 djangorestframework-inclusions==1.2.0
     # via open-api-framework
 docutils==0.21.2
-    # via django-setup-configuration
+    # via
+    #   django-markup
+    #   django-setup-configuration
+    #   python-creole
 drc-cmis==1.9.1
     # via -r requirements/base.in
 drf-extra-fields==3.4.0
@@ -303,7 +310,9 @@ jsonschema==4.7.2
 kombu==5.5.4
     # via celery
 markdown==3.3.6
-    # via -r requirements/base.in
+    # via
+    #   -r requirements/base.in
+    #   django-markup
 markupsafe==2.1.1
     # via jinja2
 maykin-2fa==1.0.1
@@ -312,6 +321,8 @@ mozilla-django-oidc==4.0.1
     # via mozilla-django-oidc-db
 mozilla-django-oidc-db==0.22.0
     # via open-api-framework
+nh3==0.3.0
+    # via textile
 notifications-api-common==0.8.1
     # via commonground-api-common
 open-api-framework==0.12.0
@@ -328,7 +339,7 @@ phonenumberslite==8.13.35
     # via django-two-factor-auth
 platformdirs==4.3.8
     # via requests-cache
-prometheus-client==0.15.0
+prometheus-client==0.22.1
     # via flower
 prompt-toolkit==3.0.31
     # via click-repl
@@ -350,6 +361,8 @@ pydantic-core==2.27.2
     # via pydantic
 pydantic-settings==2.7.0
     # via django-setup-configuration
+pygments==2.19.2
+    # via django-markup
 pyjwt==2.10.1
     # via
     #   commonground-api-common
@@ -364,6 +377,8 @@ pypng==0.20220715.0
     # via qrcode
 pyrsistent==0.18.1
     # via jsonschema
+python-creole==1.4.10
+    # via django-markup
 python-crontab==3.2.0
     # via django-celery-beat
 python-dateutil==2.8.2
@@ -383,7 +398,7 @@ python-dotenv==1.0.1
     #   pydantic-settings
 python-ipware==3.0.0
     # via django-ipware
-pytz==2022.6
+pytz==2025.2
     # via flower
 pyyaml==6.0.2
     # via
@@ -393,8 +408,10 @@ pyyaml==6.0.2
     #   zgw-consumers-oas
 qrcode==7.4.2
     # via django-two-factor-auth
-redis==4.1.4
+redis==6.2.0
     # via django-redis
+regex==2024.11.6
+    # via textile
 requests==2.32.4
     # via
     #   ape-pie
@@ -422,7 +439,6 @@ sentry-sdk==2.13.0
     # via open-api-framework
 six==1.16.0
     # via
-    #   bleach
     #   click-repl
     #   cmislib-maykin
     #   django-db-logger
@@ -430,11 +446,15 @@ six==1.16.0
     #   orderedmultidict
     #   python-dateutil
     #   url-normalize
+smartypants==2.0.2
+    # via django-markup
 sqlparse==0.5.0
     # via django
 structlog==25.4.0
     # via django-structlog
-tornado==6.5
+textile==4.0.3
+    # via django-markup
+tornado==6.5.1
     # via flower
 typing-extensions==4.12.2
     # via
@@ -464,7 +484,7 @@ urllib3==2.5.0
     #   requests
     #   requests-cache
     #   sentry-sdk
-uwsgi==2.0.26
+uwsgi==2.0.30
     # via open-api-framework
 vine==5.1.0
     # via
@@ -478,9 +498,7 @@ webauthn==2.2.0
 webencodings==0.5.1
     # via bleach
 wrapt==1.16.0
-    # via
-    #   deprecated
-    #   elastic-apm
+    # via elastic-apm
 zgw-consumers==0.37.0
     # via
     #   commonground-api-common
