+++ requirements/dev.in
@@ -13,3 +13,11 @@ django-extensions
 django-silk
 whitenoise
 pyinstrument
+
+# docs
+django-extensions
+pygraphviz
+
+# Deployment
+
+-r ../deployment/requirements.in
