+++ requirements/test-tools.in
@@ -30,6 +30,5 @@ sphinxcontrib-qthelp==1.0.2
 sphinxcontrib-serializinghtml==2.0.0
 sphinx-tabs==3.4.7
 
-
 # SPDX header checks
 click
